requirements-dev.txt
# Development and testing dependencies
# Include production dependencies (with PostgreSQL)
-r requirements.txt

# Testing frameworks
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.3.0
coverage==7.3.2

# Code quality and formatting
flake8==6.1.0
black==23.9.1